Move TestCliUpdateNotifier to shared TestServices folder
Consolidate duplicate ICliUpdateNotifier test implementations from
RunCommandTests and UpdateCommandTests into a single shared
TestCliUpdateNotifier in TestServices, combining tracking and
callback capabilities.
